+++ rich/style.py
@@ -1,4 +1,3 @@
-
 from binascii import crc32
 from functools import lru_cache
 import sys
@@ -6,7 +5,6 @@ from typing import Any, Dict, Iterable, List, Mapping, Optional, Type, Union
 
 from . import errors
 from .color import blend_rgb, Color, ColorParseError, ColorSystem
-from ._style_table import STYLE_TABLE
 from .terminal_theme import TerminalTheme, DEFAULT_TERMINAL_THEME
 
 # Style instances and style definitions are often interchangable
@@ -44,6 +42,10 @@ class Style:
         reverse (bool, optional): Enabled reverse text. Defaults to None.
         conceal (bool, optional): Enable concealed text. Defaults to None.
         strike (bool, optional): Enable strikethrough text. Defaults to None.
+        underline2 (bool, optional): Enable doubly underlined text. Defaults to None.
+        frame (bool, optional): Enable framed text. Defaults to None.
+        encircle (bool, optional): Enable encircled text. Defaults to None.
+        overline (bool, optional): Enable overlined text. Defaults to None.
         link (str, link): Link URL. Defaults to None.
     
     """
@@ -53,7 +55,32 @@ class Style:
     _attributes: int
     _set_attributes: int
 
-    __slots__ = ["_color", "_bgcolor", "_attributes", "_set_attributes", "link"]
+    __slots__ = [
+        "_color",
+        "_bgcolor",
+        "_attributes",
+        "_set_attributes",
+        "_link",
+        "_ansi",
+        "_style_definition",
+    ]
+
+    # maps bits on to SGR parameter
+    _style_map = {
+        0: "1",
+        1: "2",
+        2: "3",
+        3: "4",
+        4: "5",
+        5: "6",
+        6: "7",
+        7: "8",
+        8: "9",
+        9: "21",
+        10: "51",
+        11: "52",
+        12: "53",
+    }
 
     def __init__(
         self,
@@ -69,8 +96,15 @@ class Style:
         reverse: bool = None,
         conceal: bool = None,
         strike: bool = None,
+        underline2: bool = None,
+        frame: bool = None,
+        encircle: bool = None,
+        overline: bool = None,
         link: str = None,
     ):
+        self._ansi: Optional[str] = None
+        self._style_definition: Optional[str] = None
+
         def _make_color(color: Union[Color, str]) -> Color:
             return color if isinstance(color, Color) else Color.parse(color)
 
@@ -86,6 +120,10 @@ class Style:
             | (reverse or 0) << 6
             | (conceal or 0) << 7
             | (strike or 0) << 8
+            | (underline2 or 0) << 9
+            | (frame or 0) << 10
+            | (encircle or 0) << 11
+            | (overline or 0) << 12
         )
         self._set_attributes = (
             (bold is not None)
@@ -97,8 +135,12 @@ class Style:
             | (reverse is not None) << 6
             | (conceal is not None) << 7
             | (strike is not None) << 8
+            | (underline2 is not None) << 9
+            | (frame is not None) << 10
+            | (encircle is not None) << 11
+            | (overline is not None) << 12
         )
-        self.link = link
+        self._link = link
 
     bold = _Bit(0)
     dim = _Bit(1)
@@ -109,38 +151,85 @@ class Style:
     reverse = _Bit(6)
     conceal = _Bit(7)
     strike = _Bit(8)
+    underline2 = _Bit(9)
+    frame = _Bit(10)
+    encircle = _Bit(11)
+    overline = _Bit(12)
 
     def __str__(self) -> str:
         """Re-generate style definition from attributes."""
-        attributes: List[str] = []
-        append = attributes.append
-        if self.bold is not None:
-            append("bold" if self.bold else "not bold")
-        if self.dim is not None:
-            append("dim" if self.dim else "not dim")
-        if self.italic is not None:
-            append("italic" if self.italic else "not italic")
-        if self.underline is not None:
-            append("underline" if self.underline else "not underline")
-        if self.blink is not None:
-            append("blink" if self.blink else "not blink")
-        if self.blink2 is not None:
-            append("blink2" if self.blink2 else "not blink2")
-        if self.reverse is not None:
-            append("reverse" if self.reverse else "not reverse")
-        if self.conceal is not None:
-            append("conceal" if self.conceal else "not conceal")
-        if self.strike is not None:
-            append("strike" if self.strike else "not strike")
-        if self._color is not None:
-            append(self._color.name)
-        if self._bgcolor is not None:
-            append("on")
-            append(self._bgcolor.name)
-        if self.link:
-            append("link")
-            append(self.link)
-        return " ".join(attributes) or "none"
+        if self._style_definition is None:
+            attributes: List[str] = []
+            append = attributes.append
+            bits = self._set_attributes
+            if bits & 0b0000000001111:
+                if bits & 1:
+                    append("bold" if self.bold else "not bold")
+                if bits & (1 << 1):
+                    append("dim" if self.dim else "not dim")
+                if bits & (1 << 2):
+                    append("italic" if self.italic else "not italic")
+                if bits & (1 << 3):
+                    append("underline" if self.underline else "not underline")
+            if bits & 0b0000111110000:
+                if bits & (1 << 4):
+                    append("blink" if self.blink else "not blink")
+                if bits & (1 << 5):
+                    append("blink2" if self.blink2 else "not blink2")
+                if bits & (1 << 6):
+                    append("reverse" if self.reverse else "not reverse")
+                if bits & (1 << 7):
+                    append("conceal" if self.conceal else "not conceal")
+                if bits & (1 << 8):
+                    append("strike" if self.strike else "not strike")
+            if bits & 0b1111000000000:
+                if bits & (1 << 9):
+                    append("underline2" if self.underline2 else "not underline2")
+                if bits & (1 << 10):
+                    append("frame" if self.frame else "not frame")
+                if bits & (1 << 11):
+                    append("encircle" if self.encircle else "not encircle")
+                if bits & (1 << 12):
+                    append("overline" if self.overline else "not overline")
+            if self._color is not None:
+                append(self._color.name)
+            if self._bgcolor is not None:
+                append("on")
+                append(self._bgcolor.name)
+            if self._link:
+                append("link")
+                append(self._link)
+            self._style_definition = " ".join(attributes) or "none"
+        return self._style_definition
+
+    def _make_ansi_codes(self, color_system: ColorSystem) -> str:
+        if self._ansi is None:
+            sgr: List[str] = []
+            append = sgr.append
+            _style_map = self._style_map
+            attributes = self._attributes & self._set_attributes
+            if attributes & 0b0000000001111:
+                for bit in range(4):
+                    if attributes & (1 << bit):
+                        append(_style_map[bit])
+            if attributes & 0b0000111110000:
+                for bit in range(4, 9):
+                    if attributes & (1 << bit):
+                        append(_style_map[bit])
+            if attributes & 0b1111000000000:
+                for bit in range(9, 13):
+                    if attributes & (1 << bit):
+                        append(_style_map[bit])
+            if self._color is not None:
+                sgr.extend(self._color.downgrade(color_system).get_ansi_codes())
+            if self._bgcolor is not None:
+                sgr.extend(
+                    self._bgcolor.downgrade(color_system).get_ansi_codes(
+                        foreground=False
+                    )
+                )
+            self._ansi = ";".join(sgr)
+        return self._ansi
 
     @classmethod
     @lru_cache(maxsize=1024)
@@ -179,7 +268,7 @@ class Style:
             and self._bgcolor == other._bgcolor
             and self._set_attributes == other._set_attributes
             and self._attributes == other._attributes
-            and self.link == other.link
+            and self._link == other._link
         )
 
     def __hash__(self) -> int:
@@ -189,7 +278,7 @@ class Style:
                 self._bgcolor,
                 self._attributes,
                 self._set_attributes,
-                self.link,
+                self._link,
             )
         )
 
@@ -202,6 +291,11 @@ class Style:
     def bgcolor(self) -> Optional[Color]:
         """The background color or None if it is not set."""
         return self._bgcolor
+
+    @property
+    def link(self) -> Optional[str]:
+        """Link text, if set."""
+        return self._link
 
     @classmethod
     @lru_cache(maxsize=1024)
@@ -230,7 +324,13 @@ class Style:
             "conceal": "conceal",
             "c": "conceal",
             "strike": "strike",
-            "s": "strike"
+            "s": "strike",
+            "underline2": "underline2",
+            "uu": "underline2",
+            "frame": "frame",
+            "encircle": "encircle",
+            "overline": "overline",
+            "o": "overline",
         }
         color: Optional[str] = None
         bgcolor: Optional[str] = None
@@ -258,7 +358,7 @@ class Style:
                 if attribute is None:
                     raise errors.StyleSyntaxError(
                         f"expected style attribute after 'not', found {original_word!r}"
-                    )                
+                    )
                 attributes[attribute] = False
 
             elif word == "link":
@@ -313,6 +413,8 @@ class Style:
             append("text-decoration: underline")
         if self.strike:
             append("text-decoration: line-through")
+        if self.overline:
+            append("text-decoration: overline")
         return "; ".join(css)
 
     @classmethod
@@ -326,27 +428,19 @@ class Style:
             Style: A new style instance.
         """
 
-        style = Style()
-        update = style._update
-        for _style in styles:
-            update(_style)
-        return style
+        return sum(styles, start=Style())
 
     @classmethod
     def chain(self, *styles: "Style") -> "Style":
         """Combine styles from positional argument in to a single style.
         
         Args:
-            styles (Iterable[Style]): Styles to combine.
+            *styles (Iterable[Style]): Styles to combine.
         
         Returns:
             Style: A new style instance.
         """
-        style = Style()
-        update = style._update
-        for _style in styles:
-            update(_style)
-        return style
+        return sum(styles, start=Style())
 
     def copy(self) -> "Style":
         """Get a copy of this style.
@@ -355,11 +449,13 @@ class Style:
             Style: A new Style instance with identical attributes.
         """
         style = self.__new__(Style)
+        style._ansi = self._ansi
+        style._style_definition = self._style_definition
         style._color = self._color
         style._bgcolor = self._bgcolor
         style._attributes = self._attributes
         style._set_attributes = self._set_attributes
-        style.link = self.link
+        style._link = self._link
         return style
 
     def render(
@@ -367,7 +463,6 @@ class Style:
         text: str = "",
         *,
         color_system: Optional[ColorSystem] = ColorSystem.TRUECOLOR,
-        __STYLE_TABLE=STYLE_TABLE,
     ) -> str:
         """Render the ANSI codes for the style.
         
@@ -378,20 +473,12 @@ class Style:
         Returns:
             str: A string containing ANSI style codes.
         """
-        # Hyperlink spec: https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda
         if not text or color_system is None:
             return text
-        _attributes = self._attributes & self._set_attributes
-        attrs = [__STYLE_TABLE[_attributes]] if _attributes else []
-        if self._color is not None:
-            attrs.extend(self._color.downgrade(color_system).get_ansi_codes())
-        if self._bgcolor is not None:
-            attrs.extend(
-                self._bgcolor.downgrade(color_system).get_ansi_codes(foreground=False)
-            )
-        rendered = f"\x1b[{';'.join(attrs)}m{text}\x1b[0m" if attrs else text
-        if self.link:
-            rendered = f"\x1b]8;id={crc32(self.link.encode('utf-8'))};{self.link}\x1b\\{rendered}\x1b]8;;\x1b\\"
+        attrs = self._make_ansi_codes(color_system)
+        rendered = f"\x1b[{attrs}m{text}\x1b[0m" if attrs else text
+        if self._link:
+            rendered = f"\x1b]8;id={crc32(self._link.encode('utf-8'))};{self._link}\x1b\\{rendered}\x1b]8;;\x1b\\"
         return rendered
 
     def test(self, text: Optional[str] = None) -> None:
@@ -408,55 +495,22 @@ class Style:
         text = text or str(self)
         sys.stdout.write(f"{self.render(text)}\n")
 
-    def _apply(self, style: "Style") -> "Style":
-        """Merge this style with another.
-        
-        Args:
-            style (Optional[Style]): A style object to copy attributes from. 
-                If `style` is `None`, then a copy of this style will be returned.
-        
-        Returns:
-            (Style): A new style with combined attributes.
-
-        """
+    def __add__(self, style: Optional["Style"]) -> "Style":
+        if style is None:
+            return self
+        if not isinstance(style, Style):
+            return NotImplemented  # type: ignore
         new_style = self.__new__(Style)
+        new_style._ansi = None
+        new_style._style_definition = None
         new_style._color = style._color or self._color
         new_style._bgcolor = style._bgcolor or self._bgcolor
         new_style._attributes = (self._attributes & ~style._set_attributes) | (
             style._attributes & style._set_attributes
         )
         new_style._set_attributes = self._set_attributes | style._set_attributes
-        new_style.link = style.link or self.link
+        new_style._link = style._link or self._link
         return new_style
-
-    def _update(self, style: "Style") -> None:
-        """Update this style with another.
-        
-        Args:
-            style (Style): Style to combine to this instance.
-        """
-        self._color = style._color or self._color
-        self._bgcolor = style._bgcolor or self._bgcolor
-        self._attributes = (self._attributes & ~style._set_attributes) | (
-            style._attributes & style._set_attributes
-        )
-        self._set_attributes = self._set_attributes | style._set_attributes
-        self.link = style.link or self.link
-
-    def __add__(self, style: Optional["Style"]) -> "Style":
-        if style is None:
-            return self
-        if not isinstance(style, Style):
-            return NotImplemented  # type: ignore
-        return self._apply(style)
-
-    def __iadd__(self, style: Optional["Style"]) -> "Style":
-        if style is None:
-            return self
-        if not isinstance(style, Style):
-            return NotImplemented  # type: ignore
-        self._update(style)
-        return self
 
 
 class StyleStack: