+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - Unreleased
+
+### Added
+
+- Support for double underline, framed, encircled, and overlined attributes
+
+### Changed
+
+- Optimized Style
+
 ## [1.1.9] - 2020-05-20
 
 ### Fixed